requirements.txt
ffmpeg-python==0.1.16
pylint==2.1.1
pytest==3.7.2
pytest-parallel==0.0.2
scipy==1.1.0
tensorflow==1.10.0
